runtime/vm: extract ops, program, error into separate files
Pure code motion. vm.go (9319 lines) drops by ~570 to 8751:
- ops.go (297 lines): Op type, opcode constants, Op.String.
- program.go (216 lines): Instr, Function, Program, closure, Disassemble.
- error.go (70 lines): VMError + formatting helpers.

vm.go still holds the VM struct, Run, the eval loop (call), and the
compiler. Further splits tracked separately. Closes the first slice of
CLEAN-10.
