+++ runtime/vm/vm.go
@@ -91,508 +91,6 @@ const smallJoinThreshold = 0
 const maxCallDepth = 1048576
 
 // Op defines a VM instruction opcode.
-type Op uint8
-
-const (
-	OpConst Op = iota
-	OpMove
-	OpAdd
-	OpSub
-	OpMul
-	OpDiv
-	OpMod
-	OpEqual
-	OpNotEqual
-	OpLess
-	OpLessEq
-	OpIn
-	OpJump
-	OpJumpIfFalse
-	OpLen
-	OpIndex
-	OpSlice
-	OpSetIndex
-	OpMakeList
-	OpMakeMap
-	OpPrint
-	OpPrint2
-	OpPrintN
-	OpCall2
-	OpCall
-	OpCallV
-	OpReturn
-	OpNot
-	OpJumpIfTrue
-	OpNow
-	OpRealNow
-	OpMem
-	OpJSON
-	OpAppend
-	OpStr
-	OpUpper
-	OpLower
-	OpReverse
-	OpPadStart
-	OpInput
-	OpFirst
-	OpCount
-	OpExists
-	OpSHA256
-	OpAvg
-	OpSum
-	OpMin
-	OpMax
-	OpValues
-	OpCollect
-	OpCast
-	OpIterPrep
-	OpLoad
-	OpSave
-	OpEval
-	OpFetch
-
-	// Global variable operations
-	OpGetGlobal
-	OpSetGlobal
-
-	// Closure creation
-	OpMakeClosure
-
-	// Specialized numeric ops
-	OpAddInt
-	OpAddFloat
-	OpSubInt
-	OpSubFloat
-	OpMulInt
-	OpMulFloat
-	OpDivInt
-	OpDivFloat
-	OpModInt
-	OpModFloat
-	OpPow
-	OpRound
-	OpEqualInt
-	OpEqualFloat
-	OpLessInt
-	OpLessFloat
-	OpLessEqInt
-	OpLessEqFloat
-
-	// Unary numeric ops
-	OpNeg
-	OpNegInt
-	OpNegFloat
-
-	// List operations
-	OpUnionAll
-	OpUnion
-	OpExcept
-	OpIntersect
-	OpSort
-	OpDistinct
-	OpExpect
-	OpSelect
-
-	// Foreign function interface operations
-	OpGoCall
-	OpGoAutoCall
-	OpPyCall
-)
-
-func (op Op) String() string {
-	switch op {
-	case OpConst:
-		return "Const"
-	case OpMove:
-		return "Move"
-	case OpAdd:
-		return "Add"
-	case OpSub:
-		return "Sub"
-	case OpMul:
-		return "Mul"
-	case OpDiv:
-		return "Div"
-	case OpMod:
-		return "Mod"
-	case OpEqual:
-		return "Equal"
-	case OpNotEqual:
-		return "NotEqual"
-	case OpLess:
-		return "Less"
-	case OpLessEq:
-		return "LessEq"
-	case OpIn:
-		return "In"
-	case OpJump:
-		return "Jump"
-	case OpJumpIfFalse:
-		return "JumpIfFalse"
-	case OpLen:
-		return "Len"
-	case OpIndex:
-		return "Index"
-	case OpSlice:
-		return "Slice"
-	case OpSetIndex:
-		return "SetIndex"
-	case OpMakeList:
-		return "MakeList"
-	case OpMakeMap:
-		return "MakeMap"
-	case OpPrint:
-		return "Print"
-	case OpPrint2:
-		return "Print2"
-	case OpPrintN:
-		return "PrintN"
-	case OpCall2:
-		return "Call2"
-	case OpCall:
-		return "Call"
-	case OpCallV:
-		return "CallV"
-	case OpReturn:
-		return "Return"
-	case OpNot:
-		return "Not"
-	case OpJumpIfTrue:
-		return "JumpIfTrue"
-	case OpNow:
-		return "Now"
-	case OpRealNow:
-		return "RealNow"
-	case OpMem:
-		return "Mem"
-	case OpJSON:
-		return "JSON"
-	case OpAppend:
-		return "Append"
-	case OpStr:
-		return "Str"
-	case OpUpper:
-		return "Upper"
-	case OpLower:
-		return "Lower"
-	case OpReverse:
-		return "Reverse"
-	case OpPadStart:
-		return "PadStart"
-	case OpInput:
-		return "Input"
-	case OpFirst:
-		return "First"
-	case OpCount:
-		return "Count"
-	case OpExists:
-		return "Exists"
-	case OpSHA256:
-		return "SHA256"
-	case OpAvg:
-		return "Avg"
-	case OpSum:
-		return "Sum"
-	case OpMin:
-		return "Min"
-	case OpMax:
-		return "Max"
-	case OpValues:
-		return "Values"
-	case OpCollect:
-		return "Collect"
-	case OpCast:
-		return "Cast"
-	case OpIterPrep:
-		return "IterPrep"
-	case OpLoad:
-		return "Load"
-	case OpSave:
-		return "Save"
-	case OpEval:
-		return "Eval"
-	case OpFetch:
-		return "Fetch"
-	case OpGetGlobal:
-		return "GetGlobal"
-	case OpSetGlobal:
-		return "SetGlobal"
-	case OpMakeClosure:
-		return "MakeClosure"
-	case OpAddInt:
-		return "AddInt"
-	case OpAddFloat:
-		return "AddFloat"
-	case OpSubInt:
-		return "SubInt"
-	case OpSubFloat:
-		return "SubFloat"
-	case OpMulInt:
-		return "MulInt"
-	case OpMulFloat:
-		return "MulFloat"
-	case OpDivInt:
-		return "DivInt"
-	case OpDivFloat:
-		return "DivFloat"
-	case OpModInt:
-		return "ModInt"
-	case OpModFloat:
-		return "ModFloat"
-	case OpPow:
-		return "Pow"
-	case OpEqualInt:
-		return "EqualInt"
-	case OpEqualFloat:
-		return "EqualFloat"
-	case OpLessInt:
-		return "LessInt"
-	case OpLessFloat:
-		return "LessFloat"
-	case OpLessEqInt:
-		return "LessEqInt"
-	case OpLessEqFloat:
-		return "LessEqFloat"
-	case OpNeg:
-		return "Neg"
-	case OpNegInt:
-		return "NegInt"
-	case OpNegFloat:
-		return "NegFloat"
-	case OpUnionAll:
-		return "UnionAll"
-	case OpUnion:
-		return "Union"
-	case OpExcept:
-		return "Except"
-	case OpIntersect:
-		return "Intersect"
-	case OpSort:
-		return "Sort"
-	case OpDistinct:
-		return "Distinct"
-	case OpExpect:
-		return "Expect"
-	case OpSelect:
-		return "Select"
-	case OpGoCall:
-		return "GoCall"
-	case OpGoAutoCall:
-		return "GoAutoCall"
-	case OpPyCall:
-		return "PyCall"
-	default:
-		return "?"
-	}
-}
-
-type Instr struct {
-	Op   Op
-	A    int
-	B    int
-	C    int
-	D    int
-	Val  Value // inline constant for OpConst
-	Line int   // source line for disassembly
-}
-
-type Function struct {
-	Code       []Instr
-	NumRegs    int
-	NumParams  int
-	Name       string
-	TypeParams []string
-	Line       int // source line of function definition
-}
-
-type Program struct {
-	Funcs      []Function
-	Types      []types.Type
-	File       string   `json:"file,omitempty"`
-	Source     []string `json:"source,omitempty"`
-	NumGlobals int      `json:"num_globals,omitempty"`
-}
-
-type closure struct {
-	fn   int
-	args []Value
-}
-
-func (p *Program) funcName(idx int) string {
-	if idx < 0 || idx >= len(p.Funcs) {
-		return fmt.Sprintf("%d", idx)
-	}
-	fn := p.Funcs[idx]
-	name := fn.Name
-	if name == "" {
-		if idx == 0 {
-			return "main"
-		}
-		return fmt.Sprintf("fn%d", idx)
-	}
-	return name + formatTypeParams(fn.TypeParams)
-}
-
-// Disassemble returns a human-readable listing of the program instructions.
-// If src is provided, source lines are included as comments.
-func (p *Program) Disassemble(src string) string {
-	if src == "" && len(p.Source) > 0 {
-		src = strings.Join(p.Source, "\n")
-	}
-	lines := strings.Split(src, "\n")
-	var b strings.Builder
-	for idx, fn := range p.Funcs {
-		name := fn.Name
-		if name == "" {
-			if idx == 0 {
-				name = "main"
-			} else {
-				name = fmt.Sprintf("fn%d", idx)
-			}
-		}
-
-		// build label map for jump targets
-		labels := map[int]string{}
-		nextLabel := 0
-		for i, ins := range fn.Code {
-			switch ins.Op {
-			case OpJump:
-				if _, ok := labels[ins.A]; !ok {
-					labels[ins.A] = fmt.Sprintf("L%d", nextLabel)
-					nextLabel++
-				}
-			case OpJumpIfFalse, OpJumpIfTrue:
-				if _, ok := labels[ins.B]; !ok {
-					labels[ins.B] = fmt.Sprintf("L%d", nextLabel)
-					nextLabel++
-				}
-			}
-			_ = i
-		}
-
-		if fn.Line > 0 && fn.Line <= len(lines) {
-			fmt.Fprintf(&b, "  // %s\n", strings.TrimSpace(lines[fn.Line-1]))
-		}
-		fmt.Fprintf(&b, "func %s%s (regs=%d)\n", name, formatTypeParams(fn.TypeParams), fn.NumRegs)
-		lastLine := 0
-		for pc, ins := range fn.Code {
-			if lbl, ok := labels[pc]; ok {
-				fmt.Fprintf(&b, "%s:\n", lbl)
-			}
-			if ins.Line != lastLine && ins.Line > 0 && ins.Line <= len(lines) {
-				fmt.Fprintf(&b, "  // %s\n", strings.TrimSpace(lines[ins.Line-1]))
-				lastLine = ins.Line
-			}
-			fmt.Fprintf(&b, "  %-12s ", ins.Op)
-			switch ins.Op {
-			case OpConst:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), valueToString(ins.Val))
-			case OpMove:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpAdd, OpSub, OpMul, OpDiv, OpMod,
-				OpAddInt, OpAddFloat, OpSubInt, OpSubFloat,
-				OpMulInt, OpMulFloat, OpDivInt, OpDivFloat,
-				OpModInt, OpModFloat, OpPow,
-				OpEqual, OpNotEqual, OpEqualInt, OpEqualFloat,
-				OpLess, OpLessEq, OpLessInt, OpLessFloat,
-				OpLessEqInt, OpLessEqFloat, OpIn:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C))
-			case OpSetIndex:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C))
-			case OpJump:
-				if lbl, ok := labels[ins.A]; ok {
-					fmt.Fprintf(&b, "%s", lbl)
-				} else {
-					fmt.Fprintf(&b, "%d", ins.A)
-				}
-			case OpJumpIfFalse, OpJumpIfTrue:
-				if lbl, ok := labels[ins.B]; ok {
-					fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), lbl)
-				} else {
-					fmt.Fprintf(&b, "%s, %d", formatReg(ins.A), ins.B)
-				}
-			case OpLen:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpIndex:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C))
-			case OpSlice:
-				fmt.Fprintf(&b, "%s, %s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C), formatReg(ins.D))
-			case OpMakeList:
-				fmt.Fprintf(&b, "%s, %d, %s", formatReg(ins.A), ins.B, formatReg(ins.C))
-			case OpMakeMap:
-				fmt.Fprintf(&b, "%s, %d, %s", formatReg(ins.A), ins.B, formatReg(ins.C))
-			case OpPrint:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpPrint2:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpPrintN:
-				fmt.Fprintf(&b, "%s, %d, %s", formatReg(ins.A), ins.B, formatReg(ins.C))
-			case OpNow:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpJSON:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpAppend:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C))
-			case OpUnionAll, OpUnion, OpExcept, OpIntersect:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C))
-			case OpSort:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpStr:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpUpper:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpLower:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpReverse:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpPadStart:
-				fmt.Fprintf(&b, "%s, %s, %s, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C), formatReg(ins.D))
-			case OpInput:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpFirst:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpIterPrep:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpCount:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpExists:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpCast:
-				typ := p.Types[ins.C]
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), formatReg(ins.B), typ)
-			case OpAvg, OpSum, OpMin, OpMax:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			case OpExpect:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpMakeClosure:
-				fmt.Fprintf(&b, "%s, %s, %d, %s", formatReg(ins.A), p.funcName(ins.B), ins.C, formatReg(ins.D))
-			case OpCall2:
-				fmt.Fprintf(&b, "%s, %s, %s, %s", formatReg(ins.A), p.funcName(ins.B), formatReg(ins.C), formatReg(ins.D))
-			case OpCall:
-				fmt.Fprintf(&b, "%s, %s, %s", formatReg(ins.A), p.funcName(ins.B), formatRegs(ins.D, ins.C))
-			case OpCallV:
-				fmt.Fprintf(&b, "%s, %s, %d, %s", formatReg(ins.A), formatReg(ins.B), ins.C, formatReg(ins.D))
-			case OpGoCall, OpGoAutoCall:
-				fmt.Fprintf(&b, "%s, %s, %d, %s", formatReg(ins.A), formatReg(ins.B), ins.C, formatReg(ins.D))
-			case OpPyCall:
-				fmt.Fprintf(&b, "%s, %s, %s, %d, %s", formatReg(ins.A), formatReg(ins.B), formatReg(ins.C), ins.Val.Int, formatReg(ins.D))
-			case OpReturn:
-				fmt.Fprintf(&b, "%s", formatReg(ins.A))
-			case OpNot, OpNeg, OpNegInt, OpNegFloat:
-				fmt.Fprintf(&b, "%s, %s", formatReg(ins.A), formatReg(ins.B))
-			default:
-				fmt.Fprintf(&b, "%d,%d,%d,%d", ins.A, ins.B, ins.C, ins.D)
-			}
-			b.WriteByte('\n')
-		}
-		b.WriteByte('\n')
-	}
-	// Normalize trailing newline so the disassembly matches
-	// checked-in golden files for dataset queries.
-	out := strings.TrimRight(b.String(), "\n")
-	return out + "\n"
-}
-
 type VM struct {
 	prog    *Program
 	writer  io.Writer
@@ -607,72 +105,6 @@ type StackFrame struct {
 }
 
 // VMError wraps an error with Mochi stack trace information.
-type VMError struct {
-	Err   error
-	Stack []StackFrame
-}
-
-func (e *VMError) Error() string {
-	return e.Err.Error()
-}
-
-// Format returns a detailed representation of the VM error including the call
-// graph and annotated stack trace when program information is available.
-func (e *VMError) Format(prog *Program) string {
-	var b strings.Builder
-	b.WriteString(e.Err.Error())
-	if prog != nil {
-		b.WriteString("\ncall graph: ")
-		b.WriteString(e.callGraph())
-		b.WriteString("\nstack trace:\n")
-		b.WriteString(e.stackTrace(prog))
-	}
-	return b.String()
-}
-
-func (e *VMError) callGraph() string {
-	names := make([]string, len(e.Stack))
-	for i, f := range e.Stack {
-		names[i] = f.Func
-	}
-	return strings.Join(names, " -> ")
-}
-
-func (e *VMError) stackTrace(prog *Program) string {
-	var b strings.Builder
-	for i := len(e.Stack) - 1; i >= 0; i-- {
-		f := e.Stack[i]
-		lineInfo := ""
-		if prog != nil && f.Line > 0 && f.Line <= len(prog.Source) {
-			lineInfo = strings.TrimSpace(prog.Source[f.Line-1])
-		}
-		if lineInfo != "" {
-			fmt.Fprintf(&b, "  %s at %s:%d\n    %s\n", f.Func, prog.File, f.Line, lineInfo)
-		} else {
-			fmt.Fprintf(&b, "  %s:%d\n", f.Func, f.Line)
-		}
-	}
-	return b.String()
-}
-
-func copyTrace(trace []StackFrame) []StackFrame {
-	out := make([]StackFrame, len(trace))
-	copy(out, trace)
-	return out
-}
-
-func (m *VM) newError(err error, trace []StackFrame, line int) *VMError {
-	tr := copyTrace(trace)
-	if len(tr) > 0 {
-		tr[len(tr)-1].Line = line
-	}
-	if vmErr, ok := err.(*VMError); ok {
-		vmErr.Stack = tr
-		return vmErr
-	}
-	return &VMError{Err: err, Stack: tr}
-}
-
 func New(prog *Program, w io.Writer) *VM {
 	return &VM{prog: prog, writer: w, reader: bufio.NewReader(os.Stdin), globals: make([]Value, prog.NumGlobals)}
 }
